pyproject.toml
[tool.poetry]
name = "cdxp-campaign-service"
version = "0.0.1"
description = ""
authors = ["kafai <lamkafai1997@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
uvicorn = { extras = ["standard"], version = "^0.15.0" }
sqlmodel = "^0.0.4"
SQLAlchemy = { extras = [
  "postgresql_psycopg2binary",
  "postgresql_asyncpg",
], version = "^1.4.26" }
greenlet = "^1.1.2"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
email-validator = "^1.1.3"
databricks-cli = "^0.16.2"
APScheduler = "^3.8.1"
pandas = "^1.3.4"
adlfs = "^2021.10.0"
pyarrow = "^6.0.1"
python-dotenv = "^0.20.0"
dask = "^2022.4.1"
pytz = "^2022.1"
transitions = "^0.8.11"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pre-commit = "^2.15.0"
commitizen = "^2.20.0"
pytest = "^6.2.5"
typer = "^0.4.0"
mimesis = "^5.0.0"
jsf = "^0.3.2"

[tool.black]
line-length = 98

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
